Refractor code for rpdb

+++ src/main/java/io/github/alexiscomete/lapinousecond/worlds/WorldWithCoos.java
@@ -7,12 +7,12 @@ public class WorldWithCoos extends World {
     }
 
     @Override
-    public int getPriceForDistance(int distance, boolean place) {
+    public double getPriceForDistance(double distance, boolean place) {
         return 0;
     }
 
     @Override
-    public int getDistance(Place place1, Place place2) {
+    public double getDistance(Place place1, Place place2) {
         return 0;
     }
 }
